Исправление по замечаниям
