+++ .gitignore
@@ -14,4 +14,5 @@ node_modules/
 # Папка с собранными файлами проекта
 build/
 
-webpack.config.js
+bundle.js
+bundle.js.map
