+++ src/event.js
@@ -1,12 +1,6 @@
-import {getRandomInRange} from './main.js';
+import {Price, getRandomInRange} from './utils/index.js';
 
 export const boardEvents = document.querySelector(`.trip-day__items`);
-const Num = {
-  MIN_PRICE_EVENT: 20,
-  MAX_PRICE_EVENT: 100,
-  MIN_PRICE_SERVICE: 10,
-  MAX_PRICE_SERVICE: 200
-};
 
 const getHtmlEvent = () => {
   return `<article class="trip-point">
@@ -16,13 +10,13 @@ const getHtmlEvent = () => {
               <span class="trip-point__timetable">10:00&nbsp;&mdash; 11:00</span>
               <span class="trip-point__duration">1h 30m</span>
             </p>
-            <p class="trip-point__price">&euro;&nbsp;${getRandomInRange(Num.MIN_PRICE_EVENT, Num.MAX_PRICE_EVENT)}</p>
+            <p class="trip-point__price">&euro;&nbsp;${getRandomInRange(Price.MIN_PRICE_EVENT, Price.MAX_PRICE_EVENT)}</p>
             <ul class="trip-point__offers">
               <li>
-                <button class="trip-point__offer">Order UBER +&euro;&nbsp;${getRandomInRange(Num.MIN_PRICE_SERVICE, Num.MAX_PRICE_SERVICE)}</button>
+                <button class="trip-point__offer">Order UBER +&euro;&nbsp;${getRandomInRange(Price.MIN_PRICE_SERVICE, Price.MAX_PRICE_SERVICE)}</button>
               </li>
               <li>
-                <button class="trip-point__offer">Upgrade to business +&euro;&nbsp;${getRandomInRange(Num.MIN_PRICE_SERVICE, Num.MAX_PRICE_SERVICE)}</button>
+                <button class="trip-point__offer">Upgrade to business +&euro;&nbsp;${getRandomInRange(Price.MIN_PRICE_SERVICE, Price.MAX_PRICE_SERVICE)}</button>
               </li>
             </ul>
           </article>`;
